chore: 패키지 이동
- pieces 패키지 -> chess.pieces 패키지로 이동
- Board, ChessBoard, Rank 클래스 -> chess.board 패키지로 이동
- View, ConsoleView -> chessgame.view 패키지로 이동
