+++ src/main/java/kr/codesqaud/chessgame/chess/board/Rank.java
@@ -1,21 +1,21 @@
-package kr.codesqaud.chessgame.chess;
+package kr.codesqaud.chessgame.chess.board;
 
-import static kr.codesqaud.chessgame.pieces.Position.createPosition;
+import static kr.codesqaud.chessgame.chess.pieces.Position.createPosition;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import kr.codesqaud.chessgame.pieces.Bishop;
-import kr.codesqaud.chessgame.pieces.Blank;
-import kr.codesqaud.chessgame.pieces.King;
-import kr.codesqaud.chessgame.pieces.Knight;
-import kr.codesqaud.chessgame.pieces.Pawn;
-import kr.codesqaud.chessgame.pieces.Piece;
-import kr.codesqaud.chessgame.pieces.Queen;
-import kr.codesqaud.chessgame.pieces.Rook;
-import kr.codesqaud.chessgame.pieces.config.Color;
-import kr.codesqaud.chessgame.pieces.config.Type;
+import kr.codesqaud.chessgame.chess.pieces.Bishop;
+import kr.codesqaud.chessgame.chess.pieces.Blank;
+import kr.codesqaud.chessgame.chess.pieces.King;
+import kr.codesqaud.chessgame.chess.pieces.Knight;
+import kr.codesqaud.chessgame.chess.pieces.Pawn;
+import kr.codesqaud.chessgame.chess.pieces.Piece;
+import kr.codesqaud.chessgame.chess.pieces.Queen;
+import kr.codesqaud.chessgame.chess.pieces.Rook;
+import kr.codesqaud.chessgame.chess.pieces.config.Color;
+import kr.codesqaud.chessgame.chess.pieces.config.Type;
 
 public class Rank implements Comparable<Rank> {
