+++ src/main/java/kr/codesqaud/chessgame/chess/pieces/King.java
@@ -1,26 +1,22 @@
-package kr.codesqaud.chessgame.pieces;
+package kr.codesqaud.chessgame.chess.pieces;
 
-import static kr.codesqaud.chessgame.pieces.config.Color.BLACK;
-import static kr.codesqaud.chessgame.pieces.config.Color.WHITE;
-import static kr.codesqaud.chessgame.pieces.config.Direction.everyDirection;
-import static kr.codesqaud.chessgame.pieces.config.Type.KING;
-
+import kr.codesqaud.chessgame.chess.pieces.config.Color;
+import kr.codesqaud.chessgame.chess.pieces.config.Direction;
+import kr.codesqaud.chessgame.chess.pieces.config.Type;
 import kr.codesqaud.chessgame.exception.InvalidMovingPieceException;
-import kr.codesqaud.chessgame.pieces.config.Color;
-import kr.codesqaud.chessgame.pieces.config.Direction;
 
 public class King extends Piece {
 
     public King(final Color color, final Position position) {
-        super(color, KING, position, everyDirection());
+        super(color, Type.KING, position, Direction.everyDirection());
     }
 
     public static King createWhite(final Position position) {
-        return new King(WHITE, position);
+        return new King(Color.WHITE, position);
     }
 
     public static King createBlack(final Position position) {
-        return new King(BLACK, position);
+        return new King(Color.BLACK, position);
     }
 
     @Override
