+++ src/main/java/kr/codesqaud/chessgame/chess/pieces/Knight.java
@@ -1,14 +1,13 @@
-package kr.codesqaud.chessgame.pieces;
+package kr.codesqaud.chessgame.chess.pieces;
 
-import static kr.codesqaud.chessgame.pieces.config.Direction.knightDirection;
-import static kr.codesqaud.chessgame.pieces.config.Type.KNIGHT;
-
-import kr.codesqaud.chessgame.pieces.config.Color;
+import kr.codesqaud.chessgame.chess.pieces.config.Color;
+import kr.codesqaud.chessgame.chess.pieces.config.Direction;
+import kr.codesqaud.chessgame.chess.pieces.config.Type;
 
 public class Knight extends Piece {
 
     public Knight(final Color color, final Position position) {
-        super(color, KNIGHT, position, knightDirection());
+        super(color, Type.KNIGHT, position, Direction.knightDirection());
     }
 
     public static Knight createWhite(final Position position) {
