+++ src/main/java/kr/codesqaud/chessgame/chess/pieces/config/Color.java
@@ -1,0 +1,6 @@
+package kr.codesqaud.chessgame.chess.pieces.config;
+
+public enum Color {
+    WHITE, BLACK, NOCOLOR;
+}
+
