+++ src/main/java/kr/codesqaud/chessgame/chess/pieces/config/Type.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.chessgame.pieces.config;
+package kr.codesqaud.chessgame.chess.pieces.config;
 
 public enum Type {
     PAWN("p", 1.0),
