--- src/main/java/kr/codesqaud/chessgame/pieces/config/Color.java
@@ -1,6 +1,0 @@
-package kr.codesqaud.chessgame.pieces.config;
-
-public enum Color {
-    WHITE, BLACK, NOCOLOR;
-}
-
